+++ src/main/java/net/silentchaos512/gear/block/compounder/RefabricatorScreen.java
@@ -1,0 +1,10 @@
+package net.silentchaos512.gear.block.compounder;
+
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.util.text.ITextComponent;
+
+public class RefabricatorScreen extends CompounderScreen {
+    public RefabricatorScreen(CompounderContainer screenContainer, PlayerInventory inv, ITextComponent titleIn) {
+        super(screenContainer, inv, titleIn);
+    }
+}