+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Refabricator, a compound maker that processes cloth, fiber, and slime materials [#344]
+
 ## [2.6.19] - 2021-04-11
 ### Changed
 - Starlight charger model (naj77)