+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.6.20] - 2021-04-21
 ### Added
 - Refabricator, a compound maker that processes cloth, fiber, and slime materials [#344]
 - New models for most machine/crafting blocks (naj77)